Added !next command
Cleans up the timestamp filtering - there was no need to sanitize it for this case.

Next assumes the list of games provided is sorted - did not test yet, if that's not the case we should fix elsewhere.

Added it to an embed.

time_now is more precise so it works if you have multiple games today.