Update pride games
For some reason the pride league teams are not listed on
the website, so change the scraper to use the schedule
page. This way might be better if we can cache this for
every team after pulling it once BUT... that requires
some work.

Disables the old way of doing it.

Signed-off-by: Jarrett Schultz <jschultz38@gatech.edu>