Add !sus
Add !sus which returns all suspensions for a player.

In order to get this to work in a reasonable amount of time,
I had to cache all previous suspensions which right now is a
manual process using saveAllKHLSeasons.py. Every time a new
season is created, this would have to be updated. If I really care,
I would make a way to automatically detect on startup if a new season
has been created. So mark that one up in the issues column

Signed-off-by: Jarrett Schultz <jschultz38@gatech.edu>