Disable !sus
It still works at the moment, but I more so just want to see if anyone notices.
If no one does, I'll prob just disable it or remove it entirely. For a command
that would only be used every once in a while, it takes up way too much energy
in all aspects.

Signed-off-by: Jarrett Schultz <jschultz38@gatech.edu>